Switch RTS cabal file / package conf to use Rts.h not Stg.h
When we give cabal a configure script, it seems to begin checking
whether or not Stg.h is valid, and then gets tripped up on all the
register stuff which evidentally requires obscure command line flags to
go.

We can side-step this by making the test header Rts.h instead, which is
more normal.

I was a bit sketched out making this change, as I don't know why the
Cabal library would suddenly beging checking the header. But I did
confirm even without my RTS configure script the header doesn't compile
stand-alone, and also the Stg.h is a probably-arbitrary choice since it
dates all the way back to 2002 in
2cc5b907318f97e19b28b2ad8ed9ff8c1f401dcc.