gitlab-ci: Set HOME to plausible but still non-existent location
We have been seeing numerous CI failures on aarch64/Darwin of the form:

    CI_COMMIT_BRANCH:
    CI_PROJECT_PATH: ghc/ghc
    error: creating directory '/nonexistent': Read-only file system

Clearly *something* is attempting to create `$HOME`. A bit of sleuthing
by @int-e found that the culprit is likely `nix`, although it's not
clear why. For now we avoid the issue by setting `HOME` to a fresh
directory in the working tree.