Better wrapper activation calculation
As #20709 showed, GHC could prioritise a wrapper over a SPEC
rule, which is potentially very bad.  This patch fixes that
problem.

The fix is is described in Note [Wrapper activation], especially
item 4, 4a, and Conclusion.

For now, it has a temporary hack (replicating what was there before
to make sure that wrappers inline no earlier than phase 2.  But
it should be temporary; see #19001.