Change isUnliftedTyCon to marshalablePrimTyCon (#20401)
isUnliftedTyCon was used in three places: Ticky, Template Haskell
and FFI checks.

It was straightforward to remove it from Ticky and Template Haskell.

It is now used in FFI only and renamed to marshalablePrimTyCon.

Previously, it was fetching information from a field
in PrimTyCon called is_unlifted. Instead, I've changed the code
to compute liftedness based on the kind.

isFFITy and legalFFITyCon are removed. They were only referred from
an old comment that I removed.

There were three functions to define a PrimTyCon, but the only difference
was that they were setting is_unlifted to True or False.
Everything is now done in mkPrimTyCon.

I also added missing integer types in Ticky.hs, I think it was an oversight.

Fixes #20401