Use file hash cache when hashing object file dependencies
This fixes the immediate problem that we hash the same file multiple
different times which causes quite a noticeably performance regression.

In the future we can probably do better than this by storing the
implementation hash in the interface file rather than dependending on
hashing the object file.

Related to #20604 which notes some inefficiencies with the current
recompilation logic.

Closes #20790

-------------------------
Metric Decrease:
    T14052Type
-------------------------