iserv: Remove network dependent parts of libiserv
As noted in #20794 the parts of libiserv and iserv-proxy depend on
network, therefore are never built nor tested during CI.

Due to this iserv-proxy had bitrotted due to the bound on bytestring
being out of date.

Given we don't test this code it seems undesirable to distribute it.
Therefore, it's removed and an external maintainer can be responsible
for testing it (via head.hackage if desired).

Fixes #20794