ci: Use correct metrics baseline
It turns out there was already a function in the CI script to correctly
set the baseline for performance tests but it was just never called. I
now call it during the initialisation to set the correct baseline.

I also made the make testsuite driver take into account the
PERF_BASELINE_COMMIT environment variable

Fixes #20811