Fix panic trying to -ddump-parsed-ast for implicit fixity
A declaration such as

    infixr ++++

is supplied with an implicit fixity of 9 in the parser, but uses
an invalid SrcSpan to capture this. Use of this span triggers a panic.

Fix the problem by not recording an exact print annotation for the
non-existent fixity source.

Closes #20846

(cherry picked from commit 09b6cb45505c2c32ddaffcdb930fb3f7873b2cfc)