Properly filter for module visibility in resolvePackageImport
This completes the fix for #20779 / !7123.

Beforehand, the program worked by accident because the two versions of
the library happened to be ordered properly (due to how the hashes were
computed). In the real world I observed them being the other way around
which meant the final lookup failed because we weren't filtering for
visibility.

I modified the test so that it failed (and it's fixed by this patch).