ci: Add ENABLE_NUMA flag to explicitly turn on libnuma dependency
In recent releases a libnuma dependency has snuck into our bindists
because the images have started to contain libnuma. We now explicitly
pass `--disable-numa` to configure unless explicitly told not to by
using the `ENABLE_NUMA` environment variable.

So this is tested, there is one random validate job which builds with
--enable-numa so that the code in the RTS is still built.

Fixes #20957 and #15444