rts/PEi386: Fix handling of overflowing data relocations
As noted in #20978, the linker would previously handle overflowed
relocations by creating a trampoline. While this is fine in the case of
code symbols, it's very much not okay in the case of data symbols. To
fix this we keep track of whether each symbol is code or data and
relocate them appropriately.

Fixes #20978.