Refine tcSemigroupWarnings to work in ghc-prim
ghc-prim doesn't depend on base so can't have any Monoid or Semigroup
instances. However, attempting to load these definitions ran into issues
when the interface for `GHC.Base` did exist as that would try and load
the interface for `GHC.Types` (which is the module we are trying to
compile and has no interface).

The fix is to just not do this check when we are compiling a module in
ghc-prim.

Fixes #21069