Use a more up-to-date snapshot of the current rules in the simplifier
As the prescient (now deleted) note warns in simplifyPgmIO we have to be a bit careful
about when we gather rules from the EPS so that we get the rules for
imported bindings.

```
  -- Get any new rules, and extend the rule base
  -- See Note [Overall plumbing for rules] in GHC.Core.Rules
  -- We need to do this regularly, because simplification can
  -- poke on IdInfo thunks, which in turn brings in new rules
  -- behind the scenes.  Otherwise there's a danger we'll simply
  -- miss the rules for Ids hidden inside imported inlinings
```

Given the previous commit, the loading of unfoldings is now even more
delayed so we need to be more careful to read the EPS rule base closer to the point
where we decide to try rules.

Without this fix GHC performance regressed by a noticeably amount
because the `zip` rule was not brought into scope eagerly enough which
led to a further series of unfortunate events in the simplifer which
tipped `substTyWithCoVars` over the edge of the size threshold, stopped
it being inlined and increased allocations by 10% in some cases.

Furthermore, this change is noticeably in the testsuite as it changes
T19790 so that the `length` rules from GHC.List fires earlier.

-------------------------
Metric Increase:
    T9961
-------------------------