Draft: Use HsExpanded for untyped TH expression splices
This changes the way that untyped Template Haskell expression splices are
handled such that they are folded into the `HsExpanded` machinery. This, in
turn, makes `splitHsApps` look through TH splices, which fixes #21077.

TODO RGS: Add a more detailed description of the changes in the commit message,
and add more documentation.